Removed Service Layer and Styled Pagination
