+++ src/main/resources/templates/owners/ownersList.html
@@ -28,32 +28,34 @@
   </tr>
   </tbody>
 </table>
-
 <div th:if="${totalPages > 1}">
-  <div>
-    <div>
-      Total Rows: [[${totalItems}]]
-    </div>
-    <div>
-                    <span th:each="i: ${#numbers.sequence(1, totalPages)}">
+  <span>Pages:</span>
+  <span>[</span>
+  <span th:each="i: ${#numbers.sequence(1, totalPages)}">
       <a th:if="${currentPage != i}" th:href="@{'/ownerpage/' + ${i}}">[[${i}]]</a>
-      <span th:unless="${currentPage != i}">[[${i}]]</span> &nbsp; &nbsp;
-                    </span>
-    </div>
-    <div>
-      <a th:if="${currentPage < totalPages}" th:href="@{'/ownerpage/' + ${currentPage + 1}}">Next</a>
-      <span th:unless="${currentPage < totalPages}">Next</span>
-    </div>
-
-    <div>
-      <a th:if="${currentPage < totalPages}" th:href="@{'/ownerpage/' + ${totalPages}}">Last</a>
-      <span th:unless="${currentPage < totalPages}">Last</span>
-    </div>
-  </div>
+      <span th:unless="${currentPage != i}">[[${i}]]</span>
+    </span>
+  <span>]&nbsp;</span>
+  <span>
+      <a th:if="${currentPage > 1}" th:href="@{'/ownerpage/1'}" title="First" class="glyphicon glyphicon-backward"></a>
+      <span th:unless="${currentPage > 1}" title="First" class="glyphicon glyphicon-backward"></span>
+    </span>
+  <span>
+      <a th:if="${currentPage > 1}" th:href="@{'/ownerpage/' + ${currentPage - 1}}" title="Previous"
+         class="glyphicon glyphicon-triangle-left"></a>
+      <span th:unless="${currentPage > 1}" title="Previous" class="glyphicon glyphicon-triangle-left"></span>
+    </span>
+  <span>
+      <a th:if="${currentPage < totalPages}" th:href="@{'/ownerpage/' + ${currentPage + 1}}" title="Next"
+         class="glyphicon glyphicon-triangle-right"></a>
+      <span th:unless="${currentPage < totalPages}" title="Next" class="glyphicon glyphicon-triangle-right"></span>
+    </span>
+  <span>
+      <a th:if="${currentPage < totalPages}" th:href="@{'/ownerpage/' + ${totalPages}}" title="Last"
+         class="glyphicon glyphicon-forward"></a>
+      <span th:unless="${currentPage < totalPages}" title="Last" class="glyphicon glyphicon-forward"></span>
+    </span>
 </div>
-</div>
-
-
 </body>
 </html>
