+++ src/main/resources/templates/vets/vetList.html
@@ -18,34 +18,39 @@
   <tr th:each="vet : ${listVets}">
     <td th:text="${vet.firstName + ' ' + vet.lastName}"></td>
     <td><span th:each="specialty : ${vet.specialties}"
-              th:text="${specialty.name + ' '}" /> <span
+              th:text="${specialty.name + ' '}"/> <span
       th:if="${vet.nrOfSpecialties == 0}">none</span></td>
   </tr>
   </tbody>
 </table>
 
 <div th:if="${totalPages > 1}">
-  <div class="row col-sm-10">
-    <div class="col-sm-2">
-      Total Rows: [[${totalItems}]]
-    </div>
-    <div class="col-sm-1">
-                    <span th:each="i: ${#numbers.sequence(1, totalPages)}">
+  <span>Pages:</span>
+  <span>[</span>
+  <span th:each="i: ${#numbers.sequence(1, totalPages)}">
       <a th:if="${currentPage != i}" th:href="@{'/vetpage/' + ${i}}">[[${i}]]</a>
-      <span th:unless="${currentPage != i}">[[${i}]]</span> &nbsp; &nbsp;
-                    </span>
-    </div>
-    <div class="col-sm-1">
-      <a th:if="${currentPage < totalPages}" th:href="@{'/vetpage/' + ${currentPage + 1}}">Next</a>
-      <span th:unless="${currentPage < totalPages}">Next</span>
-    </div>
-
-    <div class="col-sm-1">
-      <a th:if="${currentPage < totalPages}" th:href="@{'/vetpage/' + ${totalPages}}">Last</a>
-      <span th:unless="${currentPage < totalPages}">Last</span>
-    </div>
-  </div>
-</div>
+      <span th:unless="${currentPage != i}">[[${i}]]</span>
+    </span>
+  <span>]&nbsp;</span>
+  <span>
+      <a th:if="${currentPage > 1}" th:href="@{'/vetpage/1'}" title="First" class="glyphicon glyphicon-backward"></a>
+      <span th:unless="${currentPage > 1}" title="First" class="glyphicon glyphicon-backward"></span>
+    </span>
+  <span>
+      <a th:if="${currentPage > 1}" th:href="@{'/vetpage/' + ${currentPage - 1}}" title="Previous"
+         class="glyphicon glyphicon-triangle-left"></a>
+      <span th:unless="${currentPage > 1}" title="Previous" class="glyphicon glyphicon-triangle-left"></span>
+    </span>
+  <span>
+      <a th:if="${currentPage < totalPages}" th:href="@{'/vetpage/' + ${currentPage + 1}}" title="Next"
+         class="glyphicon glyphicon-triangle-right"></a>
+      <span th:unless="${currentPage < totalPages}" title="Next" class="glyphicon glyphicon-triangle-right"></span>
+    </span>
+  <span>
+      <a th:if="${currentPage < totalPages}" th:href="@{'/vetpage/' + ${totalPages}}" title="Last"
+         class="glyphicon glyphicon-forward"></a>
+      <span th:unless="${currentPage < totalPages}" title="Last" class="glyphicon glyphicon-forward"></span>
+    </span>
 </div>
 </body>
 </html>
